[UPDATE]  changed data type in accountDTO for integer id to long id. Coincidence with model account. And, remove method findAll.

+++ account/src/main/java/com/banquito/core/account/dto/AccountDTO.java
@@ -8,8 +8,8 @@ import lombok.Value;
 @Builder
 public class AccountDTO {
 
-    private Integer id;
-    private Integer clientId;
+    private Long id;
+    private Long clientId;
     private String codeUniqueAccount;
     private String codeInternalAccount;
     private String codeInternationalAccount;
@@ -18,4 +18,5 @@ public class AccountDTO {
     private BigDecimal currentBalance;
     private BigDecimal availableBalance;
     private BigDecimal blockedBalance;
+    
 }
